+++ R/get_les.R
@@ -4,8 +4,6 @@
 #' [Legislative Effectiveness Scores data](https://thelawmakers.org/data-download)
 #' from the Center for Effective Lawmaking.
 #'
-#' @inheritParams get_voteview_members
-#'
 #' @param chamber Which chamber to get data for. Options are:
 #'  * `"house"`, `"h"`, `"hr"`: House data only.
 #'  * `"senate"`, `"s"`, `"sen"`: Senate data only.
@@ -34,7 +32,9 @@
 #'  that become law. LES 2.0 is only available for the 117th Congress.
 #'  Classic LES is available for the 93rd through 117th Congresses.
 #'
-#' @param local `r lifecycle::badge('experimental')` `r doc_arg_local("Center for Effective Lawmaking")`
+#' @param local_path `r lifecycle::badge('experimental')` (Optional) A file path
+#'  for reading from a local file. If no `local_path` is specified, will read
+#'  data from the Center for Effective Lawmaking website.
 #'
 #' @returns A [tibble()].
 #'
@@ -60,59 +60,53 @@
 #' # LES 2.0 (117th Congress)
 #' get_les("house", les_2 = TRUE)
 #' get_les("senate", les_2 = TRUE)
-get_les <- function(chamber, les_2 = FALSE, local = TRUE, local_dir = ".") {
-  # TODO: pass a sheet_type instead of les_2?
-  full_path <- build_file_path(data_source = "les", chamber = chamber, sheet_type = les_2,
-                               local = local, local_dir = local_dir)
-
-  # check that online connection is working
-  # TODO: fuller error handling with `get_online_data()`
-  if (R.utils::isUrl(full_path) & !crul::ok(full_path, info = F)) {
-    stop("ERROR: Could not connect to Center for Effective Lawmaking website")
+get_les <- function(chamber, les_2 = FALSE, local_path = NULL) {
+  if (is.null(local_path)) {
+    # online reading
+    # using `les_2` in place of a true `sheet_type`
+    url <- build_url(data_source = "les", chamber = chamber, sheet_type = les_2)
+    online_file <- get_online_data(url = url,
+                                   source_name = "Center for Effective Lawmaking",
+                                   return_format = "raw")
+    df <- haven::read_dta(online_file)
+  } else {
+    # local reading
+    df <- read_local_file(path = local_path, show_col_types = FALSE)
   }
 
-  haven::read_dta(full_path) |>
-    fix_les_coltypes(les_2 = les_2)
+  df <- df |>
+    # fix column types
+    fix_les_coltypes(local_path = local_path) |>
+    # convert 0/1-character `bioname` values to NA
+    dplyr::mutate(dplyr::across(.cols = "bioname",
+                                .fns = ~ dplyr::if_else(nchar(.x) <= 1, NA, .x,
+                                                        ptype = character(1))))
+
+  df
 }
 
-fix_les_coltypes <- function(df, les_2) {
+fix_les_coltypes <- function(df, local_path) {
   df <- df |>
     # using `any_of()` because of colname differences between S and HR sheets
     dplyr::mutate(dplyr::across(
-      .cols = dplyr::any_of(c("state", "st_name")),
-      as.factor)) |>
+      .cols = c("congress", "icpsr", "year", "elected",
+                "votepct", "seniority", "votepct_sq", "deleg_size",
+                "party_code", "born", "died",
+                dplyr::any_of(c("cgnum", "sensq", "thomas_num", "cd")),
+                # bill progress columns (cbill, sslaw, etc.)
+                dplyr::matches(stringr::regex("^[:lower:]{1,3}bill[12]$")),
+                dplyr::matches(stringr::regex("^[:lower:]{1,3}aic[12]$")),
+                dplyr::matches(stringr::regex("^[:lower:]{1,3}abc[12]$")),
+                dplyr::matches(stringr::regex("^[:lower:]{1,3}pass[12]$")),
+                dplyr::matches(stringr::regex("^[:lower:]{1,3}law[12]$"))),
+      .fns = as.integer)) |>
     dplyr::mutate(dplyr::across(
-      .cols = dplyr::any_of(c("congress", "cgnum", "icpsr", "year", "elected",
-                              "votepct", "seniority", "votepct_sq", "sensq",
-                              "deleg_size", "party_code", "born", "died",
-                              "thomas_num", "cd")),
-      as.integer)) |>
-    dplyr::mutate(dplyr::across(
-      .cols = dplyr::any_of(c("dem", "majority", "female", "afam", "latino",
-                              "chair", "subchr", "state_leg", "maj_leader",
-                              "min_leader", "power", "freshman", "speaker")),
-      as.logical))
+      .cols = c("dem", "majority", "female", "afam", "latino",
+                "chair", "subchr", "state_leg", "maj_leader",
+                "min_leader", "power", "freshman", dplyr::any_of("speaker")),
+      .fns = as.logical))
 
-  # bill progress columns have different names in the LES2 sheets
-  if (les_2) {
-    df <- df |>
-      dplyr::mutate(dplyr::across(
-        .cols = dplyr::any_of(c("cbill2", "caic2", "cabc2", "cpass2", "claw2",
-                                "sbill2", "saic2", "sabc2", "spass2", "slaw2",
-                                "ssbill2", "ssaic2", "ssabc2", "sspass2", "sslaw2",
-                                "expectation2", "allbill2", "allaic2",
-                                "allabc2", "allpass2", "alllaw2")),
-        as.integer))
-  } else {
-    df <- df |>
-      dplyr::mutate(dplyr::across(
-        .cols = dplyr::any_of(c("cbill1", "caic1", "cabc1", "cpass1", "claw1",
-                                "sbill1", "saic1", "sabc1", "spass1", "slaw1",
-                                "ssbill1", "ssaic1", "ssabc1", "sspass1", "sslaw1",
-                                "expectation1", "allbill1", "allaic1",
-                                "allabc1", "allpass1", "alllaw1")),
-        as.integer))
-  }
+  df <- df |> create_factor_columns(local_path = local_path)
 
   df
 }