+++ R/get_voteview_members.R
@@ -6,7 +6,7 @@
 #' Congress (i.e., each member is listed once for every two years in office).
 #'
 #' @param chamber Which chamber to get data for. Options are:
-#'  * `"all"`, `"congress"`: Both House and Senate data (the default).
+#'  * `"all"`, `"congress"`, `"hs"`: Both House and Senate data (the default).
 #'  * `"house"`, `"h"`, `"hr"`: House data only.
 #'  * `"senate"`, `"s"`, `"sen"`: Senate data only.
 #'  These options are case-insensitive. If you explicitly pass a different value,
@@ -22,10 +22,9 @@
 #'  If specified, Congress numbers cannot be greater than the [current_congress()]
 #'  (i.e., you cannot try to get future data).
 #'
-#' @param local `r lifecycle::badge('experimental')` `r doc_arg_local("Voteview")`
-#'
-#' @param local_dir `r lifecycle::badge('experimental')` The directory containing
-#'  the local file. Defaults to the working directory.
+#' @param local_path `r lifecycle::badge('experimental')` (Optional) A file path
+#'  for reading from a local file. If no `local_path` is specified, will read
+#'  data from the Voteview website.
 #'
 #' @returns A [tibble()].
 #'
@@ -39,16 +38,13 @@
 #' Please cite this dataset as:
 #'
 #' Lewis, Jeffrey B., Keith Poole, Howard Rosenthal, Adam Boche, Aaron Rudkin,
-#' and Luke Sonnet (2023). *Voteview: Congressional Roll-Call Votes Database*.
+#' and Luke Sonnet (2024). *Voteview: Congressional Roll-Call Votes Database*.
 #' <https://voteview.com/>
 #'
 #' @export
 #'
 #' @examplesIf interactive()
 #' get_voteview_members()
-#'
-#' # Force to get data from Voteview website
-#' get_voteview_members(local = FALSE)
 #'
 #' # Get data for only one chamber
 #' # NOTE: the President is included in all data
@@ -64,26 +60,54 @@
 #' # Get data for a set of Congresses
 #' get_voteview_members(congress = 1:10)
 #'
-get_voteview_members <- function(chamber = "all", congress = NULL, local = TRUE, local_dir = ".") {
-  # join multiple congresses
-  if (length(congress) > 1 & is.numeric(congress)) {
-    list_of_dfs <- lapply(congress, function(.cong) get_voteview_members(local = local,
-                                                                         local_dir = local_dir,
-                                                                         chamber = chamber,
-                                                                         congress = .cong))
+get_voteview_members <- function(chamber = "all", congress = NULL, local_path = NULL) {
+  # join multiple congresses (for online downloads)
+  if (length(congress) > 1 && is.numeric(congress) && is.null(local_path)) {
+    list_of_dfs <- lapply(congress, function(.cong) {
+      get_voteview_members(chamber = chamber,
+                           congress = .cong,
+                           local_path = local_path)
+    })
     return(dplyr::bind_rows(list_of_dfs))
   }
 
-  full_path <- build_file_path(data_source = "voteview", chamber = chamber, congress = congress,
-                               sheet_type = "members", local = local, local_dir = local_dir)
-
-  # request data from online
-  if (R.utils::isUrl(full_path)) {
-    full_path <- get_online_data(url = full_path, source_name = "Voteview")
+  if (is.null(local_path)) {
+    # online reading
+    url <- build_url(data_source = "voteview", chamber = chamber, congress = congress,
+                     sheet_type = "members")
+    online_file <- get_online_data(url = url, source_name = "Voteview")
+    df <- readr::read_csv(online_file, col_types = "ifiinfiiiccnnnnnniilnn")
+  } else {
+    # local reading
+    df <- read_local_file(path = local_path, col_types = "ifiinfiiiccnnnnnniilnn")
   }
 
-  readr::read_csv(full_path,
-                  col_types = "ifiinfiiiccnnnnnni") |>
+  if (isTRUE(tools::file_ext(local_path) == "dta")) {
+    # fixes for coming from .dta files
+    df <- df |>
+      # no need to specify levels if data is already coming from saved DTA file
+      dplyr::mutate(dplyr::across(.cols = c("chamber", "state_abbrev"),
+                                  .fns = haven::as_factor),
+                    dplyr::across(.cols = "conditional",
+                                  .fns = as.logical),
+                    dplyr::across(.cols = "bioguide_id",
+                                  .fns = ~ dplyr::na_if(.x, "")))
+  } else {
     # fix order of state abbreviations
-    dplyr::mutate(state_abbrev = factor(state_abbrev, levels = c(datasets::state.abb, "USA")))
+    df <- df |>
+      dplyr::mutate(dplyr::across(.cols = "state_abbrev",
+                                  .fns = ~ factor(.x, levels = c(datasets::state.abb, "USA"))))
+  }
+
+  if (!is.null(local_path)) {
+    df <- df |>
+      filter_congress(congress = congress) |>
+      filter_chamber(chamber = chamber)
+  }
+
+  df <- df |>
+    dplyr::mutate(dplyr::across(.cols = c("district_code", "born", "died"),
+                                .fns = as.integer))
+
+  df
 }